refactor: Clean up plugin directory after modularization
- Replace monolithic uemcp_listener.py with modular version
- Keep backup of old listener as uemcp_listener_monolithic_backup.py
- Move documentation files to docs/modularization/
- Update README.md with modular architecture information
- Remove redundant uemcp_listener_modular.py file

The plugin directory is now organized with clear module separation.
