+++ docs/modularization/MODULAR_ARCHITECTURE.md
